Add document summary section to results.html and integrate summarization in document_analysis view
